CLOUDSTACK-9993: Fixes from code review
